feat: much improved file delegate
The file delegate now shows the size of the file.

The file delegate now features an explicit download button,
which allows starting and stopping a download. When the
download is complete, there is now a button to open the file.

Download progress is now visible.

There is also now a tooltip introducing users to the fact that
an icon is drag and droppable when the file completes downloading.

Fixes #20.