feat: revamped upload dialog on desktop
Tok now has a better upload sequence on desktop, relying on only one button
to send a file instead of one per filetype.

You can caption files in the dialog, and choose whether or not
to compress an image when you send it.

This also adds support for sending video and audio files as their
own message types.